Move InvalidArgumentException to common module
