+++ backend/common/src/main/java/ai/verta/modeldb/common/exceptions/InvalidArgumentException.java
@@ -1,6 +1,5 @@
-package ai.verta.modeldb.exceptions;
+package ai.verta.modeldb.common.exceptions;
 
-import ai.verta.modeldb.common.exceptions.ModelDBException;
 import com.google.rpc.Code;
 
 public class InvalidArgumentException extends ModelDBException {
