+++ backend/src/main/java/ai/verta/modeldb/experimentRun/subtypes/FeatureHandler.java
@@ -2,10 +2,10 @@ package ai.verta.modeldb.experimentRun.subtypes;
 
 import ai.verta.modeldb.Feature;
 import ai.verta.modeldb.common.exceptions.InternalErrorException;
+import ai.verta.modeldb.common.exceptions.InvalidArgumentException;
 import ai.verta.modeldb.common.futures.FutureJdbi;
 import ai.verta.modeldb.common.futures.InternalFuture;
 import ai.verta.modeldb.common.subtypes.MapSubtypes;
-import ai.verta.modeldb.exceptions.InvalidArgumentException;
 import java.util.AbstractMap;
 import java.util.HashSet;
 import java.util.List;
